feat: manual receive address part 2
