+++ src/components/MultiHopTrade/components/TradeConfirm/ReceiveSummary.tsx
@@ -1,28 +1,14 @@
+import { ChevronDownIcon, ChevronUpIcon, QuestionIcon } from '@chakra-ui/icons'
 import {
-  CheckIcon,
-  ChevronDownIcon,
-  ChevronUpIcon,
-  CloseIcon,
-  EditIcon,
-  QuestionIcon,
-} from '@chakra-ui/icons'
-import {
-  Box,
   Collapse,
   Divider,
   Flex,
-  IconButton,
-  Input,
-  InputGroup,
-  InputRightElement,
   Skeleton,
   Stack,
-  Tooltip,
   useColorModeValue,
   useDisclosure,
 } from '@chakra-ui/react'
-import type { AssetId } from '@shapeshiftoss/caip'
-import { isLedger } from '@shapeshiftoss/hdwallet-ledger'
+import { type AssetId } from '@shapeshiftoss/caip'
 import type { AmountDisplayMeta, ProtocolFee, SwapSource } from '@shapeshiftoss/swapper'
 import { SwapperName } from '@shapeshiftoss/swapper'
 import type { PartialRecord } from '@shapeshiftoss/types'
@@ -30,20 +16,17 @@ import { type FC, memo, useCallback, useMemo, useState } from 'react'
 import { useTranslate } from 'react-polyglot'
 import { Amount } from 'components/Amount/Amount'
 import { HelperTooltip } from 'components/HelperTooltip/HelperTooltip'
-import { useReceiveAddress } from 'components/MultiHopTrade/hooks/useReceiveAddress'
 import { Row, type RowProps } from 'components/Row/Row'
 import { RawText, Text } from 'components/Text'
 import type { TextPropTypes } from 'components/Text/Text'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
-import { useFeatureFlag } from 'hooks/useFeatureFlag/useFeatureFlag'
-import { useWallet } from 'hooks/useWallet/useWallet'
 import { bnOrZero } from 'lib/bignumber/bignumber'
 import { fromBaseUnit } from 'lib/math'
 import {
   THORCHAIN_LONGTAIL_STREAMING_SWAP_SOURCE,
   THORCHAIN_STREAM_SWAP_SOURCE,
 } from 'lib/swapper/swappers/ThorchainSwapper/constants'
-import { isSome, middleEllipsis } from 'lib/utils'
+import { isSome } from 'lib/utils'
 import {
   selectActiveQuoteAffiliateBps,
   selectQuoteAffiliateFeeUserCurrency,
@@ -55,6 +38,7 @@ import {
 import { useAppSelector } from 'state/store'
 
 import { FeeModal } from '../FeeModal/FeeModal'
+import { RecipientAddress } from '../TradeInput/components/RecipientAddress'
 
 type ReceiveSummaryProps = {
   isLoading?: boolean
@@ -70,22 +54,12 @@ type ReceiveSummaryProps = {
   swapSource?: SwapSource
 } & RowProps
 
-const editIcon = <EditIcon />
-const checkIcon = <CheckIcon />
-const closeIcon = <CloseIcon />
-
 const shapeShiftFeeModalRowHover = { textDecoration: 'underline', cursor: 'pointer' }
 
 const tradeFeeSourceTranslation: TextPropTypes['translation'] = [
   'trade.tradeFeeSource',
   { tradeFeeSource: 'ShapeShift' },
 ]
-
-// TODO(gomes): implement me
-const isCustomRecipientAddress = false
-const recipientAddressTranslation: TextPropTypes['translation'] = isCustomRecipientAddress
-  ? 'trade.customRecipientAddress'
-  : 'trade.recipientAddress'
 
 export const ReceiveSummary: FC<ReceiveSummaryProps> = memo(
   ({
@@ -162,45 +136,12 @@ export const ReceiveSummary: FC<ReceiveSummaryProps> = memo(
       setShowFeeModal(!showFeeModal)
     }, [showFeeModal])
 
-    // Recipient address state and handlers
-    const [isRecipientAddressEditing, setIsRecipientAddressEditing] = useState(false)
-    const handleEditRecipientAddressClick = useCallback(() => {
-      setIsRecipientAddressEditing(true)
-    }, [])
-
-    const handleCancelClick = useCallback(() => {
-      setIsRecipientAddressEditing(false)
-    }, [])
-
-    const handleSaveClick = useCallback(() => {
-      setIsRecipientAddressEditing(false)
-    }, [])
-
-    const handleInputChange: React.ChangeEventHandler<HTMLInputElement> = useCallback(event => {
-      // TODO(gomes): dispatch here and make the input controlled with a local state value and form context so that
-      // typing actually does something
-      // dispatch(tradeInput.actions.setManualReceiveAddress(undefined))
-      console.log(event.target.value)
-    }, [])
+    // Recipient address logic, form, state and handlers
 
     const minAmountAfterSlippageTranslation: TextPropTypes['translation'] = useMemo(
       () => ['trade.minAmountAfterSlippage', { slippage: slippageAsPercentageString }],
       [slippageAsPercentageString],
     )
-
-    const wallet = useWallet().state.wallet
-    const useReceiveAddressArgs = useMemo(
-      () => ({
-        fetchUnchainedAddress: Boolean(wallet && isLedger(wallet)),
-      }),
-      [wallet],
-    )
-    const isHolisticRecipientAddressEnabled = useFeatureFlag('HolisticRecipientAddress')
-
-    const receiveAddress = useReceiveAddress(useReceiveAddressArgs)
-
-    // This should never happen but it may
-    if (isHolisticRecipientAddressEnabled && !receiveAddress) return null
 
     return (
       <>
@@ -351,77 +292,7 @@ export const ReceiveSummary: FC<ReceiveSummaryProps> = memo(
                   </Row>
                 </>
               )}
-
-            {/* TODO(gomes): This should probably be made its own component and <ManualAddressEntry /> removed */}
-            {/* TODO(gomes): we can safely remove this condition when this feature goes live */}
-            {isHolisticRecipientAddressEnabled &&
-              receiveAddress &&
-              (isRecipientAddressEditing ? (
-                <InputGroup size='sm'>
-                  <Input
-                    value={''} // TODO: Controlled input value
-                    onChange={handleInputChange}
-                    autoFocus
-                    placeholder={translate('trade.customRecipientAddressDescription')}
-                  />
-                  <InputRightElement width='4.5rem'>
-                    <Box
-                      display='flex'
-                      alignItems='center'
-                      justifyContent='space-between'
-                      width='full'
-                      px='2'
-                    >
-                      <Box
-                        as='button'
-                        display='flex'
-                        alignItems='center'
-                        justifyContent='center'
-                        borderRadius='md'
-                        onClick={handleSaveClick}
-                      >
-                        {checkIcon}
-                      </Box>
-                      <Box
-                        as='button'
-                        display='flex'
-                        alignItems='center'
-                        justifyContent='center'
-                        borderRadius='md'
-                        onClick={handleCancelClick}
-                      >
-                        {closeIcon}
-                      </Box>
-                    </Box>
-                  </InputRightElement>
-                </InputGroup>
-              ) : (
-                <>
-                  <Divider borderColor='border.base' />
-                  <Row>
-                    <Row.Label>
-                      <Text translation={recipientAddressTranslation} />
-                    </Row.Label>
-                    <Row.Value whiteSpace='nowrap'>
-                      <Stack direction='row' spacing={1} alignItems='center'>
-                        <RawText>{middleEllipsis(receiveAddress)}</RawText>
-                        <Tooltip
-                          label={translate('trade.customRecipientAddressDescription')}
-                          placement='top'
-                          hasArrow
-                        >
-                          <IconButton
-                            aria-label='Edit recipient address'
-                            icon={editIcon}
-                            variant='ghost'
-                            onClick={handleEditRecipientAddressClick}
-                          />
-                        </Tooltip>
-                      </Stack>
-                    </Row.Value>
-                  </Row>
-                </>
-              ))}
+            <RecipientAddress />
           </Stack>
         </Collapse>
         <FeeModal isOpen={showFeeModal} onClose={handleFeeModal} />
